+++ packages/server/src/routes.ts
@@ -2,19 +2,29 @@ import { Express, Request, Response } from 'express';
 
 import { getRTData } from 'solax-common/services';
 
-import { ServerDb } from './db';
+import { ServerDb } from './server-db';
 
 export const current = (app: Express) =>
   app.get('/current', async (req: Request, res: Response) =>
     res.send(await getRTData())
   );
 
-export const thisMinute = (app: Express, db: ServerDb) =>
+export const getThisMinute = (app: Express, db: ServerDb) =>
   app.get('/minute/now', async (req: Request, res: Response) =>
     res.send(await db.getMinute())
   );
 
-export const allDays = (app: Express, db: ServerDb) =>
-  app.get('/day/all', async (req: Request, res: Response) =>
+export const getAllDays = (app: Express, db: ServerDb) =>
+  app.get('/day', async (req: Request, res: Response) =>
     res.send(await db.getAllDays())
+  );
+
+export const getDay = (app: Express, db: ServerDb) =>
+  app.get('/day/:date', async ({ params }: Request, res: Response) =>
+    res.send(await db.getDay(new Date(params.date)))
+  );
+
+export const getDayMinutes = (app: Express, db: ServerDb) =>
+  app.get('/day/:date/minute', async ({ params }: Request, res: Response) =>
+    res.send(await db.getDayMinutes(new Date(params.date)))
   );
